Добавлен софт удаление программ тренировок и тренировок если есть ссылка на них в сессии. если нет ссылки то хард удаление

+++ app/src/main/java/ru/squidory/trainingdairymobile/ui/trainings/WorkoutManagementAdapter.java
@@ -89,7 +89,7 @@ public class WorkoutManagementAdapter extends RecyclerView.Adapter<WorkoutManage
             });
 
             deleteButton.setOnClickListener(v -> {
-                if (listener != null) {
+                if (listener != null && !Boolean.TRUE.equals(workout.getIsDeleted())) {
                     listener.onDeleteWorkout(workout);
                 }
             });
